feat: Add sequece function to monads/free/utils.ts
This commit adds a new function called `sequece` to the `monads/free/utils.ts` file. The `sequece` function takes in multiple `Free<Program<A>>` instances and returns a `Free<Program<A[]>>` instance. It uses the `reduce` and `map` methods to chain the programs together and collect the results into an array. This function can be used to sequence and execute multiple programs in a specific order.

Based on the recent user commits and repository commits, it seems that there have been updates to package versions and dependencies for the functional-ts monads and http packages. Additionally, there have been refactors to the updates and ReaderT monad chain methods.